% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{base_simulator}
\alias{base_simulator}
\title{base simulator}
\usage{
base_simulator(
  N_Policies,
  policy_exprs,
  policy_parameters,
  frequency_exprs,
  frequnecy_parameters,
  severity_exprs,
  severity_paramters
)
}
\arguments{
\item{N_Policies}{number of policies}

\item{policy_exprs}{alist of policy attribute formula, not NULL}

\item{policy_parameters}{alist of parameters of policy formula}

\item{frequency_exprs}{alist of frequency attribute formula, the first formula should describe claims count, not NULL}

\item{frequnecy_parameters}{alist of parameters of frequency formula}

\item{severity_exprs}{alist of sevevrity attribute formula, the first formula should describe loss, not NULL}

\item{severity_paramters}{alist of parameters of severity formula}
}
\description{
This is the basic simulation model that doesn't not specifically define the transition of claims, such as open/close, payment, and also case reserve development
}
