% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_proportion.R
\name{cdf_proportion}
\alias{cdf_proportion}
\title{cdf propotion calc}
\usage{
cdf_proportion(q, distr = "exp", ...)
}
\arguments{
\item{q}{a vector of numbers}

\item{distr}{distribution name, for example 'lnorm" for lognormal}

\item{...}{additional distrbituion parameter of the p\link{distr} function, such as meanlog and sdlog for lnorm}
}
\description{
helper function to calculate the cdfs in between various qs
}
