% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claims transition functions.R
\name{claims_trans}
\alias{claims_trans}
\title{claims transition}
\usage{
claims_trans(df)
}
\arguments{
\item{df}{data frame after evaluating on severity transit components}
}
\description{
claims transition
}
