% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tidysimloss_obj.R
\name{crt_tidysimloss}
\alias{crt_tidysimloss}
\title{create a tidysimloss object}
\usage{
crt_tidysimloss(policy_df, policy_required_field_map, frequency_alist,
  frequency_params_components_alist, severity_init_components_alist,
  severity_transit_components_alist, severity_params_components_alist)
}
\arguments{
\item{policy_df}{Polidy Data frame}

\item{policy_required_field_map}{maps the required fields, Policy Number, Policy Effective Date, Policy Expriation Date}

\item{frequency_alist}{Frequency Component:}

\item{severity_init_components_alist}{initial State components:}

\item{severity_transit_components_alist}{transition state components:}

\item{severity_params_components_alist}{parameter components:}
}
\description{
this function creates a tidysimloss object, based on a policy listing, as well as a claims transaction list
}
