% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Truncation Functions.R
\name{dtrunc}
\alias{dtrunc}
\title{trunc d function}
\usage{
dtrunc(x, FUN, Att = 0, rTrunc = Inf, ...)
}
\arguments{
\item{FUN}{distribution}

\item{Att}{attachment}

\item{rTrunc}{right truncation}

\item{...}{additional arguments}
}
\description{
d truncation function
}
