% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_parser.R
\name{expr_evaluation}
\alias{expr_evaluation}
\title{evalute parsed expressions}
\usage{
expr_evaluation(df, expr_alist)
}
\arguments{
\item{df}{data frame}

\item{expr_alist}{an alist that contains expressions}
}
\description{
evaluate parsed expressions
}
