% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_parser.R
\name{expr_parser}
\alias{expr_parser}
\title{expression parser}
\usage{
expr_parser(expr_alist)
}
\arguments{
\item{expr_alist}{an alist that contains expressions}
}
\description{
parses expressions based on conventions specified in this package
}
