% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_claims_generator.R
\name{ini_claims_gen}
\alias{ini_claims_gen}
\alias{crt_tidysimloss}
\title{generate initial claims states}
\usage{
crt_tidysimloss(
  policy_df,
  policy_required_field_map,
  frequency_alist,
  frequency_params_components_alist,
  severity_init_components_alist,
  severity_transit_components_alist,
  severity_params_components_alist
)
}
\arguments{
\item{policy_df}{Polidy Data frame}

\item{policy_required_field_map}{maps the required fields, Policy Number, Policy Effective Date, Policy Expriation Date}

\item{severity_init_components_alist}{initial State components:}

\item{severity_params_components_alist}{parameter components:}
}
\description{
generates initial claims
}
