% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Truncation Functions.R
\name{pseudo}
\alias{pseudo}
\title{seudo p function}
\usage{
pseudo(Att, ...)
}
\arguments{
\item{Att}{attachment}

\item{...}{additional arguments}
}
\description{
this is a pseudo p function, that is used when p function not available
}
