% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Truncation Functions.R
\name{ptrunc}
\alias{ptrunc}
\title{trunc p function}
\usage{
ptrunc(x, FUN, Att = 0, rTrunc = Inf, ...)
}
\arguments{
\item{FUN}{distribution}

\item{Att}{attachment}

\item{rTrunc}{right truncation}

\item{...}{additional arguments}
}
\description{
p truncation function
}
