% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdiscrete_function.R
\name{rdiscrete}
\alias{rdiscrete}
\title{discrete sampling}
\usage{
rdiscrete(n, options)
}
\arguments{
\item{n}{number of obs}

\item{options}{options to be sampled}
}
\description{
discrete sampling
}
