% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Truncation Functions.R
\name{rtrunc}
\alias{rtrunc}
\title{trunc r function}
\usage{
rtrunc(n, FUN, Att = 0, rTrunc = Inf, ...)
}
\arguments{
\item{FUN}{distribution}

\item{Att}{attachment}

\item{rTrunc}{right truncation}

\item{...}{additional arguments}
}
\description{
r truncation function
}
