% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_result_methods.R
\name{save_sim_result}
\alias{save_sim_result}
\title{save simulation results}
\usage{
save_sim_result(con, tab, df)
}
\arguments{
\item{con}{a DBI connection}

\item{tab}{name of the table}

\item{df}{name of the dataframe}
}
\description{
save simulation results
}
